+++ fe/src/components/home-page/ArtistBanner.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import ArrowRightAltIcon from "@mui/icons-material/ArrowRightAlt";
 import { Box, Chip, Typography } from "@mui/material";
 import { Link } from "react-router-dom";
-import ArtistCarousel from "./ArtistCarousel";
-import path from "../constants/path";
+import ArtistCarousel from "../ArtistCarousel";
+import path from "../../constants/path";
 
 export default function ArtistBanner() {
   return (
