+++ fe/src/layouts/ArtistLayout.jsx
@@ -2,7 +2,7 @@
 import { Box } from "@mui/material";
 import React from "react";
 import Sidebar from "../components/Sidebar";
-import ArtistNavbar from "../components/ArtistNavbar";
+import ArtistNavbar from "../components/layout/ArtistNavbar";
 
 export default function ArtistLayout({ children }) {
   return (
