+++ fe/src/pages/Homepage.jsx
@@ -1,11 +1,11 @@
 import { Box } from "@mui/material";
 import React from "react";
-import Headers from "../components/Headers";
-import StatCard from "../components/StatCard";
-import TopServices from "../components/TopServices";
+import Headers from "../components/layout/Headers";
+import StatCard from "../components/home-page/StatCard";
+import TopServices from "../components/home-page/TopServices";
 import Offers from "../components/Offers";
-import Footer from "../components/Footer";
-import ArtistBanner from "../components/ArtistBanner";
+import Footer from "../components/layout/Footer";
+import ArtistBanner from "../components/home-page/ArtistBanner";
 
 export default function Homepage() {
   return (
